12.move inner class to upper level.
