+++ src/test/java/com/ejoongseok/wmslive/outbound/feature/MaterialType.java
@@ -1,0 +1,10 @@
+package com.ejoongseok.wmslive.outbound.feature;
+
+enum MaterialType {
+    CORRUGATED_BOX("골판지 상자");
+    private final String description;
+
+    MaterialType(final String description) {
+        this.description = description;
+    }
+}
